Treemap counting a weight of zero as one
And I added Documentation for treemap.py:

The script adds one "change" towards the changes of a file,
counting it as the "initial commit",
which are not represented in the data seen in the treemap
because file names from the older data set that are not in the newer data set will be added with a weight of "zero".
Since "plotly.express" does not show paths with a value of zero, a "initial commit" has to be added.