Adding and finishing comparison.py
The script compares the paths of the older data set with the newer data set,
by adding the paths to a list.
The script then checks for each line in the list of paths from older data set if the line is in the list of paths from the newer data set.
If not, the path will be added to the file with a weight (changes) of zero.