Adding Compare Feature und fixing new Treemap code (#7)
* Downloading pandas and plotly and rewriting treemap code 

* Treemap working with test data

The path to the test data is now integrated into the code and can thus be used for testing changes on the treemap code.

* Adding to Persona Hans

I stated that he wants clearity and a quick overview.

* Added Git Log Command DOCUMENTATION

I added two recommended git log commands that the user might use to gain data from a repo.

* Treemap counting a weight of zero as one

And I added Documentation for treemap.py:

The script adds one "change" towards the changes of a file,
counting it as the "initial commit",
which are not represented in the data seen in the treemap
because file names from the older data set that are not in the newer data set will be added with a weight of "zero".
Since "plotly.express" does not show paths with a value of zero, a "initial commit" has to be added.

* Adding and finishing comparison.py

The script compares the paths of the older data set with the newer data set,
by adding the paths to a list.
The script then checks for each line in the list of paths from older data set if the line is in the list of paths from the newer data set.
If not, the path will be added to the file with a weight (changes) of zero.

* Treemap Script now checks if the line is empty

---------

Co-authored-by: JesusIsKing