Treemap.py takes argv and creates two file paths
The Script now takes the argv and creates two file paths for the script to use later in the code.
These two paths are the paths to the two data sets created after the git command.
The user must make sure that these data sets (txt files) are in the same dir as the script.