Recoded count.py
Function is suppoed to be called in createHotspots.py
In count.py's count_lines function are two file paths takes as parameter.
They will be opened the changes for every file will be counted.
Puts the new data into two new files and appends the their filename to a list.
Returns that list.