Script can be run with one file and clarg (#15)
* Treemap.py takes argv and creates two file paths

The Script now takes the argv and creates two file paths for the script to use later in the code.
These two paths are the paths to the two data sets created after the git command.
The user must make sure that these data sets (txt files) are in the same dir as the script.

* Changed name of treemap.py to createHotspots.py

* Recoded count.py

Function is suppoed to be called in createHotspots.py
In count.py's count_lines function are two file paths takes as parameter.
They will be opened the changes for every file will be counted.
Puts the new data into two new files and appends the their filename to a list.
Returns that list.

* Changed the name to treemap.py again useless tbh ._.

* Rewritten comparison.py

Function added compare_data, takes the files from count_lines.
Creates the data for the display of the actual treemap.

* Renamed file path variables

* Refactored and added function to display treemap

* Moved functions from other python files

Making the other two python files absolete.

* compare_data now returns treemap data file path

* Writing script

* Refactored count_lines for naming consistency

* Renamed treemap.py to createHotspots.py

* Fixed the issue of files getting stored in the wrong directory

---------

Co-authored-by: JesusIsKing