Added date as clarg and tests
The Date for the git log command,
so the user of the script can individually define the "border" between the old and new data.

The format of the date is checked in a different function to be used in get_data function.
The get_data function gets the data from the repository and stores them in two textfiles automatically.

I also added tests. Y? Obvious.