Added Script func that includes caching
Also the times how long it takes to run the script.
Including the caching the time of execution is reduced by around 50%.
Caching means that the function checks if data already exists in the starting folder, if yes it uses the already existing data.
Added two functions that check if the data exits.

Refactored their tests for better understandability.