Fixed debug bug (wr permission in dir missing)
When debugging in VS Code the cwd is 'C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0'
A permission error is thrown when trying to create a new text file (i.e. when trying to create cache files)

refactored some functions like check_if_data_exists(=>data_exists) and check_date_format(=>date_format_correct)
they both now return booleans for cleaner implementation and flexibility

The script (not the function) has also been refactored; now tempered to the needs of the check for writing-permission
it now checks if the inputs of the user are correct; the refactored functions mentioned above help to make this process more flexible
then checks for wr-permissions in the cwd by trying to create an empy cache file (old.txt)
This not only helps to debug in VS Code with launch.json but also makes the system more robust
Only downside: if the current working directory has no writing-permission, cache will be created in the repo given by the user
Solution: The cacheHandler tool!